refactor Cleared whole tools file and ordered everything in proper file
